collect-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paramiko==3.3.1
pysnmp==4.4.12
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.2
pika==1.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
